+++ src/Components/DocumentDBConnection.php
@@ -2,6 +2,7 @@
 
 namespace DreamFactory\Core\Azure\Components;
 
+use DreamFactory\Core\Exceptions\RestException;
 use DreamFactory\DocumentDb\Client;
 use DreamFactory\DocumentDb\Resources\Collection;
 use DreamFactory\DocumentDb\Resources\Document;
@@ -13,6 +14,12 @@ class DocumentDBConnection
 
     /** @var string */
     protected $database;
+
+    /** @var array */
+    protected $headers = [];
+
+    /** Limit option */
+    const OPT_LIMIT = 'limit';
 
     /**
      * DocumentDBConnection constructor.
@@ -27,6 +34,45 @@ class DocumentDBConnection
         $this->database = $database;
     }
 
+    /**
+     * Set additional request headers.
+     *
+     * @param array $header DocumentDB REST API request header
+     */
+    public function setHeaders(array $header)
+    {
+        $this->headers = array_merge($this->headers, $header);
+    }
+
+    /**
+     * Sets options.
+     *
+     * @param array $options
+     */
+    public function setOptions(array $options)
+    {
+        $limit = array_get($options, static::OPT_LIMIT);
+
+        if (!empty($limit)) {
+            $this->setHeaders(['x-ms-max-item-count' => $limit]);
+        }
+    }
+
+    /**
+     * Sets request headers.
+     *
+     * @return array
+     */
+    public function getHeaders()
+    {
+        $headers = [];
+        foreach ($this->headers as $key => $value) {
+            $headers[] = $key . ': ' . $value;
+        }
+
+        return $headers;
+    }
+
     /*********************************
      * Collection operations
      *********************************/
@@ -34,12 +80,17 @@ class DocumentDBConnection
     /**
      * List all collections in a Database.
      *
+     * @param array $options Operation options
+     *
      * @return array
      */
-    public function listCollections()
+    public function listCollections(array $options = [])
     {
+        $this->setOptions($options);
         $coll = new Collection($this->client, $this->database);
+        $coll->setHeaders($this->getHeaders());
         $rs = $coll->list();
+        $this->checkResponse($rs);
         $colls = array_get($rs, 'DocumentCollections');
         $list = [];
         if (!empty($colls)) {
@@ -54,14 +105,18 @@ class DocumentDBConnection
     /**
      * Retrieves a collection information.
      *
-     * @param string $id Collection ID
+     * @param string $id      Collection ID
+     * @param array  $options Operation options
      *
      * @return array
      */
-    public function getCollection($id)
+    public function getCollection($id, array $options = [])
     {
+        $this->setOptions($options);
         $coll = new Collection($this->client, $this->database);
+        $coll->setHeaders($this->getHeaders());
         $rs = $coll->get($id);
+        $this->checkResponse($rs);
         unset($rs['_curl_info']);
 
         return $rs;
@@ -70,14 +125,18 @@ class DocumentDBConnection
     /**
      * Creates a collection.
      *
-     * @param array $data Collection data
+     * @param array $data    Collection data
+     * @param array $options Operation options
      *
      * @return array
      */
-    public function createCollection(array $data)
+    public function createCollection(array $data, array $options = [])
     {
+        $this->setOptions($options);
         $coll = new Collection($this->client, $this->database);
+        $coll->setHeaders($this->getHeaders());
         $rs = $coll->create($data);
+        $this->checkResponse($rs);
         unset($rs['_curl_info']);
 
         return $rs;
@@ -86,15 +145,19 @@ class DocumentDBConnection
     /**
      * Replaces a collection.
      *
-     * @param array  $data Collection data
+     * @param array  $data    Collection data
      * @param string $id
+     * @param array  $options Operation options
      *
      * @return array
      */
-    public function replaceCollection(array $data, $id)
+    public function replaceCollection(array $data, $id, array $options = [])
     {
+        $this->setOptions($options);
         $coll = new Collection($this->client, $this->database);
+        $coll->setHeaders($this->getHeaders());
         $rs = $coll->replace($data, $id);
+        $this->checkResponse($rs);
         unset($rs['_curl_info']);
 
         return $rs;
@@ -103,14 +166,18 @@ class DocumentDBConnection
     /**
      * Deletes a collection.
      *
-     * @param string $id Collection ID
+     * @param string $id      Collection ID
+     * @param array  $options Operation options
      *
      * @return array
      */
-    public function deleteCollection($id)
+    public function deleteCollection($id, array $options = [])
     {
+        $this->setOptions($options);
         $coll = new Collection($this->client, $this->database);
+        $coll->setHeaders($this->getHeaders());
         $rs = $coll->delete($id);
+        $this->checkResponse($rs);
         unset($rs['_curl_info']);
 
         return $rs;
@@ -124,13 +191,17 @@ class DocumentDBConnection
      * List all documents in a collection.
      *
      * @param string $collection Collection to list documents from
+     * @param array  $options    Operation options
      *
      * @return array
      */
-    public function listDocuments($collection)
+    public function listDocuments($collection, array $options = [])
     {
+        $this->setOptions($options);
         $doc = new Document($this->client, $this->database, $collection);
+        $doc->setHeaders($this->getHeaders());
         $rs = $doc->list();
+        $this->checkResponse($rs);
         unset($rs['_curl_info']);
 
         return $rs;
@@ -141,13 +212,17 @@ class DocumentDBConnection
      *
      * @param string $collection Collection to list document from
      * @param string $id         ID of the document to retrieve
+     * @param array  $options    Operation options
      *
      * @return array
      */
-    public function getDocument($collection, $id)
+    public function getDocument($collection, $id, array $options = [])
     {
+        $this->setOptions($options);
         $doc = new Document($this->client, $this->database, $collection);
+        $doc->setHeaders($this->getHeaders());
         $rs = $doc->get($id);
+        $this->checkResponse($rs);
         unset($rs['_curl_info']);
 
         return $rs;
@@ -158,13 +233,17 @@ class DocumentDBConnection
      *
      * @param string $collection Collection to create a document in
      * @param array  $data       Document data
+     * @param array  $options    Operation options
      *
      * @return array
      */
-    public function createDocument($collection, array $data)
+    public function createDocument($collection, array $data, array $options = [])
     {
+        $this->setOptions($options);
         $doc = new Document($this->client, $this->database, $collection);
+        $doc->setHeaders($this->getHeaders());
         $rs = $doc->create($data);
+        $this->checkResponse($rs);
         unset($rs['_curl_info']);
 
         return $rs;
@@ -176,13 +255,17 @@ class DocumentDBConnection
      * @param string $collection Collection to replace a document in
      * @param array  $data       Document data
      * @param string $id         ID of the document being replaced
+     * @param array  $options    Operation options
      *
      * @return array
      */
-    public function replaceDocument($collection, array $data, $id)
+    public function replaceDocument($collection, array $data, $id, array $options = [])
     {
+        $this->setOptions($options);
         $doc = new Document($this->client, $this->database, $collection);
+        $doc->setHeaders($this->getHeaders());
         $rs = $doc->replace($data, $id);
+        $this->checkResponse($rs);
         unset($rs['_curl_info']);
 
         return $rs;
@@ -191,15 +274,19 @@ class DocumentDBConnection
     /**
      * Deletes a document.
      *
-     * @param $collection Collection to delete a document from
-     * @param $id         ID of the document to delete
+     * @param string $collection Collection to delete a document from
+     * @param string $id         ID of the document to delete
+     * @param array  $options    Operation options
      *
      * @return array
      */
-    public function deleteDocument($collection, $id)
+    public function deleteDocument($collection, $id, array $options = [])
     {
+        $this->setOptions($options);
         $doc = new Document($this->client, $this->database, $collection);
+        $doc->setHeaders($this->getHeaders());
         $rs = $doc->delete($id);
+        $this->checkResponse($rs);
         unset($rs['_curl_info']);
 
         return $rs;
@@ -211,15 +298,36 @@ class DocumentDBConnection
      * @param string $collection Collection to perform the query on
      * @param string $sql        SQL query string
      * @param array  $params     Query parameters
+     * @param array  $options    Operation options
      *
      * @return array
      */
-    public function queryDocument($collection, $sql, array $params = [])
+    public function queryDocument($collection, $sql, array $params = [], array $options = [])
     {
+        $this->setOptions($options);
         $doc = new Document($this->client, $this->database, $collection);
+        $doc->setHeaders($this->getHeaders());
         $rs = $doc->query($sql, $params);
+        $this->checkResponse($rs);
         unset($rs['_curl_info']);
 
         return $rs;
+    }
+
+    /**
+     * Checks response status code for exceptions.
+     *
+     * @param array $response
+     *
+     * @throws \DreamFactory\Core\Exceptions\RestException
+     */
+    protected function checkResponse(array $response)
+    {
+        $responseCode = intval(array_get($response, '_curl_info.http_code'));
+        $message = array_get($response, 'message');
+
+        if ($responseCode >= 400) {
+            throw new RestException($responseCode, $message);
+        }
     }
 }