+++ src/Services/DocumentDB.php
@@ -115,4 +115,10 @@ class DocumentDB extends BaseNoSqlDbService
         $this->removeFromCache('table_names');
         $this->tables = [];
     }
+
+    /** {@inheritdoc} */
+    protected function handlePATCH()
+    {
+        return false;
+    }
 }